[BE] fix: 동화 참여 기록 페이지 저장시 ID 반환하도록 수정
- 본인과 상대방의 참여 기록 페이지 ID를 반환하도록 수정 완료

+++ BE/project/src/main/java/com/ssafy/project/dto/record/BookRecordPageDto.java
@@ -13,10 +13,10 @@ import lombok.NoArgsConstructor;
 public class BookRecordPageDto {
     private Long bookRecordId;
     private Long partnerBookRecordId;
-    private Long bookRecordPageNumber;
+    private int bookRecordPageNumber;
     private String pagePath;
     private String audioPath;
     private String role;
     private String text;
-    private Long audioNumber;
+    private int audioNumber;
 }
